Makefile + settings anonymizer added
To prevent the true settings.py file to be pushed upstream, a generic version of this file is needed. So as to make changes on settings.py impact the generic version of this file, a Makefile has been added. Every time the server is run (target run, runserver), the new version of settings-generic.py is built, if needed.