Settings and anonymization issues
Rather than recreating an anonymous version of settings.py every time the app is run (which artificially adds changes to settings-generic.py), everything that should be kept secret is deported to a particular file which is git-ignored so as to prevent it from being pushed.

Now we can also safely work with the real settings.py file.